python_client/requirements.txt
pyserial
pandas
matplotlib
openpyxl
